# lpc-payment-backend-api/docs/swagger.yaml
openapi: 3.0.0
info:
  description: >
    This is the API for SDK for the Loyalty Payment Card.


    All dates are in ISO 8601 format (YYYY-MM-DD). They need to be converted
    from/to MMDDYYYY when interacting with SOAP.


    Pagination for GET requests is controlled via `page` and `per_page` query
    parameteres and presented via HTTP Headers:

    - `X-Total-Count` – total number of elements, eg. "X-Total-Count: 200". Response for GET requests in `payments` group don't contain this header.

    - `X-Page` – current page, eg. "X-Page: 1"

    - `X-Per-Page` – limit of elements per page, eg. "X-Per-Page: 20"

    - `Link` – link relations to first, last, next and previous pages, eg.
    "Link: \<https://example.com/api/v1/payments/cards?page=1&per_page=20\>;
    rel=\"first\",
    \<https://example.com/api/v1/payments/cards?page=2&per_page=20\>;
    rel=\"prev\",
    \<https://example.com/api/v1/payments/cards?page=4&per_page=20\>;
    rel=\"next\",
    \<https://example.com/api/v1/payments/cards?page=10&per_page=20\>;
    rel=\"last\""
  version: 1.0.0
  title: Poseidon LPC
security:
  - jwt: []
servers:
  - url: /api/v1
...
paths:
...
  /loyalty/rules:
    get:
      tags:
        - loyaltyRules
      operationId: listRules
      summary: List Rules
      description: |
        **Authorization**: JWT, only Admin

        Get paginated data of all the rules
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Loyalty Rules received successfully
          headers:
            X-Total-Count:
              $ref: '#/components/parameters/XTotalCountParam'
            X-Page:
              $ref: '#/components/parameters/XPageParam'
            X-Per-Page:
              $ref: '#/components/parameters/XPerPageParam'
            Link:
              $ref: '#/components/parameters/LinkParam'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoyaltyRule'
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not an Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
    post:
      tags:
        - loyaltyRules
      operationId: createRule
      summary: create a Rule
      description: |
        **Authorization**: JWT, only Admin

        Create a Rule for given operation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoyaltyRuleCreate'
      responses:
        '201':
          description: Loyalty Rule created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyRule'
        '400':
          description: Invalid data in body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not an Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
  '/loalty/rules/{ruleId}':
    parameters:
      - $ref: '#/components/parameters/loyaltyRuleIdParam'
    get:
      tags:
        - loyaltyRules
      operationId: getRule
      summary: Get a Rule by ID
      description: |
        **Authorization**: JWT, only Admin

        Get a Rule by ID
      responses:
        '200':
          description: Loyalty Rule received successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyRule'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not an Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: There is no Rule with that ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
    patch:
      tags:
        - loyaltyRules
      operationId: partialUpdateRule
      summary: Partially update a Rule
      description: |
        **Authorization**: JWT, only Admin

        Partially update a Rule by ID. Only specied fields will be updated.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoyaltyRule'
      responses:
        '200':
          description: Loyalty Rule updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyRule'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not an Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: There is no Rule with that ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
    put:
      tags:
        - loyaltyRules
      operationId: updateRule
      summary: Update a Rule
      description: |
        **Authorization**: JWT, only Admin

        Update a Rule by ID. All fields are required.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoyaltyRuleCreate'
      responses:
        '200':
          description: Loyalty Rule updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyRule'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not an Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: There is no Rule with that ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
    delete:
      tags:
        - loyaltyRules
      operationId: deleteRule
      summary: Delete a Rule
      description: |
        **Authorization**: JWT, only Admin

        Delete a Rule by ID.
      responses:
        '204':
          description: Loyalty Rule deleted successfully
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not an Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: There is no Rule with that ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
  /loyalty/points:
    get:
      tags:
        - loyaltyPoints
      operationId: listTransactions
      summary: List Loyalty Points Transactions
      description: |
        **Authorization**: JWT, only Admin

        Get paginated data of all the Loyalty Points Transactions
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
      responses:
        '200':
          description: Loyalty Points Transactions received successfully
          headers:
            X-Total-Count:
              $ref: '#/components/parameters/XTotalCountParam'
            X-Page:
              $ref: '#/components/parameters/XPageParam'
            X-Per-Page:
              $ref: '#/components/parameters/XPerPageParam'
            Link:
              $ref: '#/components/parameters/LinkParam'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LoyaltyTransaction'
        '400':
          description: Invalid pagination parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not an Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
    post:
      tags:
        - loyaltyPoints
      operationId: createTransaction
      summary: create a Loyalty Points Transaction
      description: |
        **Authorization**: JWT, only Admin

        Create a Transaction for given operation
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoyaltyTransactionCreate'
      responses:
        '201':
          description: Loyalty Points Transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyTransaction'
        '400':
          description: Invalid data in body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not an Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
  '/loalty/points/{transactionId}':
    parameters:
      - $ref: '#/components/parameters/loyaltyTransactionIdParam'
    get:
      tags:
        - loyaltyPoints
      operationId: getTransaction
      summary: Get a Transaction
      description: |
        **Authorization**: JWT, only Admin

        Get a Loyalty Points Transaction by ID
      responses:
        '200':
          description: Loyalty Points Transaction received successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyTransaction'
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '403':
          description: Not an Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '404':
          description: There is no Loyalty Points Transaction with that ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFound'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
  /loyalty/points/redeem:
    put:
      tags:
        - loyaltyPoints
      operationId: redeemPoints
      summary: Redeem Loyalty Points
      description: |
        **Authorization**: JWT any logged in user

        Redeem Loyalty Points by an authenticated user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoyaltyPoints'
      responses:
        '200':
          description: Loyalty Points Transaction redeemed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyTransaction'
        '400':
          description: Invalid points
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
  /loyalty/points/buy:
    put:
      tags:
        - loyaltyPoints
      operationId: buyPoints
      summary: Buy Loyalty Points
      description: |
        **Authorization**: JWT any logged in user

        Buy Loyalty Points by an authenticated user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoyaltyPoints'
      responses:
        '200':
          description: Loyalty Points Transaction bought successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoyaltyTransaction'
        '400':
          description: Invalid points
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequest'
        '401':
          description: No JWT or JTW is invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Unauthorized'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerError'
components:
  parameters:
...
    XTotalCountParam:
      name: X-Total-Count
      in: header
      example: 100
      description: Total number of records
      schema:
        type: integer
    XPageParam:
      name: X-Page
      in: header
      example: 1
      description: Current page
      schema:
        type: integer
    XPerPageParam:
      name: X-Per-Page
      in: header
      example: 20
      description: Total number of pages
      schema:
        type: integer
    LinkParam:
      name: Link
      in: header
      example: >-
        Link: <https://example.com/api/v1/payments/cards?page=1&per_page=20>;
        rel="first",
        <https://example.com/api/v1/payments/cards?page=2&per_page=20>;
        rel="next",
        <https://example.com/api/v1/payments/cards?page=10&per_page=20>;
        rel="last"
      description: relations to other pages
      schema:
        type: string
    PageParam:
      name: page
      in: query
      example: 1
      schema:
        type: integer
        default: 1
    PerPageParam:
      name: per_page
      in: query
      example: 20
      schema:
        type: integer
        default: 20
...
    loyaltyRuleIdParam:
      name: ruleId
      in: path
      description: Rule ID
      required: true
      schema:
        type: integer
        format: int32
        example: 123
    loyaltyTransactionIdParam:
      name: transactionId
      in: path
      description: Loyalty Points Transaction ID
      required: true
      schema:
        type: integer
        format: int64
        example: 12345
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoyaltyRule:
      type: object
      properties:
        id:
          type: integer
          format: int32
          readOnly: true
          description: Rule ID
          example: 12345
        operation:
          type: string
          description: Name of operation associated with the rule
          example: make payment
        rule:
          type: object
          description: >-
            represantation of the rule in json-rules-engine form, see
            https://www.npmjs.com/package/json-rules-engine
          example:
              {
                conditions: {
                  all: [
                    {
                      fact: 'amount',
                      operator: 'greaterThanInclusive',
                      value: 100,
                    },
                    {
                      fact: 'amount',
                      operator: 'lessThanInclusive',
                      value: 200,
                    }
                  ]
                }
              }
        reward:
          type: integer
          format: int32
          description: Points awarded when the rule passes
          example: 123
    LoyaltyRuleCreate:
      required:
        - operation
        - rule
        - reward
      allOf:
        - $ref: '#/components/schemas/LoyaltyRule'
    LoyaltyTransaction:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
          description: Loyalty Points Transaction ID
          example: 123
        userId:
          type: integer
          format: int32
          description: User ID
          example: 12345
        ruleId:
          type: integer
          format: int32
          nullable: true
          default: null
          description: 'Loyalty Rule ID, can be NULL'
          example: 12345
        amount:
          type: integer
          format: int32
          description: Points added or taken
          example: -20
    LoyaltyTransactionCreate:
      required:
        - userId
        - amount
      allOf:
        - $ref: '#/components/schemas/LoyaltyTransaction'
    LoyaltyPoints:
      type: object
      required:
        - points
      properties:
        points:
          type: integer
          format: int32
          description: Points to redeem/buy
          example: 500
    BadRequest:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          description: Error message
          example: Bad request
    Unauthorized:
      type: object
      properties:
        code:
          type: integer
          example: 401
        message:
          type: string
          description: Error message
          example: Unauthorized
    Forbidden:
      type: object
      properties:
        code:
          type: integer
          example: 403
        message:
          type: string
          description: Error message
          example: Forbidden
    NotFound:
      type: object
      properties:
        code:
          type: integer
          example: 404
        message:
          type: string
          description: Error message
          example: Not found
...
    ServerError:
      type: object
      properties:
        code:
          type: integer
          example: 500
        message:
          type: string
          description: Error message
          example: Internal server error